TR-1784 - Rename Form to SettingsForm, update tests to match new implementation
