+++ src/pages/templatesV2/components/settings/SettingsSection.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Panel } from '@sparkpost/matchbox';
-import SettingsForm from './Form.Container';
+import SettingsForm from './SettingsForm.Container';
 import styles from './SettingsSection.module.scss';
 
 const SettingsSection = () => (
